+++ src/components/button/button.tsx
@@ -1,24 +1,20 @@
-import classnames from 'classnames';
+import cn from 'classnames';
 import { ComponentProps, FC } from 'react';
 
 import css from './button.module.css';
 
 type ButtonVariant = 'primary' | 'secondary';
 
-interface ButtonProps extends ComponentProps<'button'> {
+interface IButtonProps extends ComponentProps<'button'> {
 	variant?: ButtonVariant;
 }
 
-export const Button: FC<ButtonProps> = (props) => {
+export const Button: FC<IButtonProps> = (props) => {
 	const { className, ...restProps } = props;
 
 	return (
 		<button
-			className={classnames(
-				css.root,
-				css[`button_${props.variant}`],
-				className
-			)}
+			className={cn(css.root, css[`button_${props.variant}`], className)}
 			{...restProps}
 		>
 			{props.children}
