+++ src/components/button/index.ts
@@ -1,1 +1,1 @@
-export * from './button.component';
+export * from './button';
