+++ src/modules/dashboard/components/connection-list/connection-card/connection-card.tsx
