+++ src/modules/dashboard/components/connection-list/connection-list.skeleton.tsx
@@ -1,4 +1,4 @@
-import { ConnectionCardSkeleton } from '../connection-card';
+import { ConnectionCardSkeleton } from './connection-card';
 import css from './connection-list.module.css';
 
 const skeletonCount = 8;
