+++ src/modules/dashboard/components/index.ts
@@ -1,2 +1,3 @@
+export { ConnectionSummary } from './connection-summary/connection-summary';
 export { ConnectionFilters } from './connection-filters/connection-filters';
 export { ConnectionList } from './connection-list/connection-list';
