+++ src/store/clients/types.ts
@@ -1,8 +1,8 @@
-export interface Client {
+export type Client = {
 	id: string;
 	status: ClientStatus;
 	address: { ip: string; port: string };
 	hostname: string;
-}
+};
 
 export type ClientStatus = 'active' | 'inactive' | 'waiting';
